pkg: consolidated console and agent/logger into agent/console, put daemon, logging, and registry under misc
