+++ pkg/misc/registry/registry.go
