+++ pkg/misc/registry/registry_test.go
